metrics/pyproject.toml
[tool.poetry]
name = "datagov-metrics"
version = "0.1.0"
description = ""
authors = ["robert-bryson <robert.bryson@gsa.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.129.0"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
apiclient = "^1.0.4"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
